requirements.txt
fastapi
uvicorn
requests
pydantic
aiofiles
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
redis[asyncio]
psycopg2-binary
PyJWT>=2.9
python-multipart
bcrypt
cryptography
